+++ app/locations/routes.py
@@ -13,6 +13,11 @@ from app.panelists.reports.debut_by_year import retrieve_show_years
 from .reports.average_scores import retrieve_average_scores_by_location
 from .reports.home_vs_away import retrieve_location_home_vs_away
 from .reports.recordings_by_year import retrieve_recording_counts_by_year
+from .reports.show_recordings import (
+    retrieve_location,
+    retrieve_locations,
+    retrieve_recording_details,
+)
 
 blueprint = Blueprint("locations", __name__, template_folder="templates")
 
@@ -47,6 +52,68 @@ def home_vs_away() -> str:
     return render_template("locations/home-vs-away.html", show_counts=_show_counts)
 
 
+@blueprint.route("/recordings-by-year", methods=["GET", "POST"])
+def recordings_by_year() -> str:
+    """View: Recordings by Year Report."""
+    _database_connection = mysql.connector.connect(**current_app.config["database"])
+    _locations = retrieve_locations(database_connection=_database_connection)
+    _show_years = retrieve_show_years(database_connection=_database_connection)
+
+    if request.method == "POST":
+        _location = "location" in request.form and request.form["location"]
+        _year = "year" in request.form and request.form["year"]
+        try:
+            year = int(_year)
+        except ValueError:
+            year = None
+
+        if year not in _show_years:
+            _database_connection.close()
+            return render_template(
+                "locations/recordings-by-year.html",
+                locations=_locations,
+                years=_show_years,
+                recordings=None,
+            )
+
+        _location_info = retrieve_location(
+            location_slug=_location, database_connection=_database_connection
+        )
+        _recordings = retrieve_recording_details(
+            location_slug=_location,
+            year=_year,
+            database_connection=_database_connection,
+        )
+        if not _recordings:
+            _database_connection.close()
+            return render_template(
+                "locations/recordings-by-year.html",
+                locations=_locations,
+                location_info=_location_info,
+                years=_show_years,
+                recordings=None,
+            )
+
+        _database_connection.close()
+        return render_template(
+            "locations/recordings-by-year.html",
+            locations=_locations,
+            years=_show_years,
+            location_info=_location_info,
+            year=year,
+            recordings=_recordings,
+        )
+
+    # Fallback for GET request
+    _database_connection.close()
+    return render_template(
+        "locations/recordings-by-year.html",
+        locations=_locations,
+        years=_show_years,
+        recordings=None,
+    )
+
+
 @blueprint.route("/recording-counts-by-year", methods=["GET", "POST"])
 def recording_counts_by_year() -> str:
     """View: Recording Counts by Year Report."""
@@ -54,7 +121,6 @@ def recording_counts_by_year() -> str:
     _show_years = retrieve_show_years(database_connection=_database_connection)
 
     if request.method == "POST":
-        # Parse panelist dropdown selections
         _year = "year" in request.form and request.form["year"]
         try:
             year = int(_year)