+++ app/scorekeepers/routes.py
@@ -6,11 +6,14 @@
 """Scorekeepers Routes for Wait Wait Reports."""
 
 import mysql.connector
-from flask import Blueprint, current_app, render_template
+from flask import Blueprint, current_app, render_template, request
 
 from app.panelists.reports.appearances_by_year import retrieve_all_years
 
-from .reports.appearances import retrieve_appearance_summaries
+from .reports.appearances import (
+    retrieve_appearance_summaries,
+    retrieve_scorekeepers,
+)
 from .reports.appearances_by_year import (
     retrieve_all_appearance_counts,
     retrieve_all_appearance_counts_by_year,
@@ -20,6 +23,7 @@ from .reports.introductions import (
     retrieve_all_scorekeeper_introductions,
     retrieve_scorekeepers_with_introductions,
 )
+from .reports.show_apperances import retrieve_appearance_details
 
 blueprint = Blueprint("scorekeepers", __name__, template_folder="templates")
 
@@ -69,6 +73,61 @@ def appearance_counts_by_year_grid() -> str:
         "scorekeepers/appearance-counts-by-year-grid.html",
         scorekeepers=_scorekeepers,
         show_years=_show_years,
+    )
+
+
+@blueprint.route("/appearances-by-year", methods=["GET", "POST"])
+def appearances_by_year() -> str:
+    """View: Appearances by Year Report."""
+    _database_connection = mysql.connector.connect(**current_app.config["database"])
+    _scorekeepers = retrieve_scorekeepers(database_connection=_database_connection)
+    _scorekeepers_dict = {
+        scorekeeper["slug"]: scorekeeper["name"] for scorekeeper in _scorekeepers
+    }
+    _show_years = retrieve_show_years(database_connection=_database_connection)
+
+    if request.method == "POST":
+        _scorekeeper = "scorekeeper" in request.form and request.form["scorekeeper"]
+        _year = "year" in request.form and request.form["year"]
+        try:
+            _year = int(_year)
+        except ValueError:
+            _year = None
+
+        if _scorekeeper in _scorekeepers_dict and _year in _show_years:
+            _scorekeeper_info = {
+                "slug": _scorekeeper,
+                "name": _scorekeepers_dict[_scorekeeper],
+            }
+            _appearances = retrieve_appearance_details(
+                scorekeeper_slug=_scorekeeper,
+                year=_year,
+                database_connection=_database_connection,
+            )
+            _database_connection.close()
+            return render_template(
+                "scorekeepers/appearances-by-year.html",
+                scorekeepers=_scorekeepers,
+                years=_show_years,
+                scorekeeper_info=_scorekeeper_info,
+                appearances=_appearances,
+            )
+
+        # No valid host returned
+        _database_connection.close()
+        return render_template(
+            "scorekeepers/appearances-by-year.html",
+            scorekeepers=_scorekeepers,
+            years=_show_years,
+            average_scores=None,
+        )
+
+    # Fallback for GET request
+    _database_connection.close()
+    return render_template(
+        "scorekeepers/appearances-by-year.html",
+        scorekeepers=_scorekeepers,
+        years=_show_years,
     )